requirements.txt
fastapi
uvicorn[standard]
openai
pydantic
python-dotenv
numpy
faiss-cpu